{{cookiecutter.project_slug}}/{{cookiecutter.project_slug}}/requirements/dev.txt
-r base.txt
PyMySQL
django-debug-toolbar
django-extensions